+++ tests/Unit/Service/Bulk/BulkCreateUsersServiceTest.php
@@ -24,7 +24,7 @@ namespace OAT\SimpleRoster\Tests\Unit\Service\Bulk;
 
 use OAT\SimpleRoster\Entity\LineItem;
 use OAT\SimpleRoster\Lti\Service\AssigmentFactoryInterface;
-use OAT\SimpleRoster\Lti\Service\AssignmentDecorator\AssignmentCollectionDecoratorInterface;
+use OAT\SimpleRoster\Lti\Service\AssignmentCollectionMapper\AssignmentCollectionMapperInterface;
 use OAT\SimpleRoster\Lti\Service\UserGenerator\UserGeneratorStateStorageInterface;
 use OAT\SimpleRoster\Repository\LineItemRepository;
 use OAT\SimpleRoster\Service\Bulk\BulkCreateUsersService;
@@ -38,8 +38,8 @@ class BulkCreateUsersServiceTest extends TestCase
 {
     /** @var UserGeneratorStateStorageInterface|MockObject */
     private $stateStorage;
-    /** @var AssignmentCollectionDecoratorInterface|MockObject */
-    private $assignmentDecorator;
+    /** @var AssignmentCollectionMapperInterface|MockObject */
+    private $assignmentMapper;
     /** @var LineItemAssignedIndexResolver|MockObject */
     private $lineItemAssignedIndexResolver;
     /** @var AssigmentFactoryInterface|MockObject */
@@ -54,7 +54,7 @@ class BulkCreateUsersServiceTest extends TestCase
     protected function setUp(): void
     {
         $this->stateStorage = $this->createMock(UserGeneratorStateStorageInterface::class);
-        $this->assignmentDecorator = $this->createMock(AssignmentCollectionDecoratorInterface::class);
+        $this->assignmentMapper = $this->createMock(AssignmentCollectionMapperInterface::class);
         $this->lineItemAssignedIndexResolver = $this->createMock(LineItemAssignedIndexResolver::class);
         $this->assigmentFactory = $this->createMock(AssigmentFactoryInterface::class);
         $this->storage = $this->createMock(StorageInterface::class);
@@ -63,7 +63,7 @@ class BulkCreateUsersServiceTest extends TestCase
 
         $this->bulkCreateUsersService = new BulkCreateUsersService(
             $this->stateStorage,
-            $this->assignmentDecorator,
+            $this->assignmentMapper,
             $this->lineItemAssignedIndexResolver,
             $this->assigmentFactory,
             $this->storage,
